Remove AzureBlobMetadata
