+++ bag_replicator/src/test/scala/uk/ac/wellcome/platform/archive/bagreplicator/replicator/s3/S3ReplicatorTest.scala
@@ -2,14 +2,14 @@ package uk.ac.wellcome.platform.archive.bagreplicator.replicator.s3
 
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.platform.archive.bagreplicator.replicator.ReplicatorTestCases
+import uk.ac.wellcome.platform.archive.bagreplicator.replicator.models.{ReplicationRequest, ReplicationSucceeded}
+import uk.ac.wellcome.storage.Identified
 import uk.ac.wellcome.storage.fixtures.S3Fixtures
 import uk.ac.wellcome.storage.fixtures.S3Fixtures.Bucket
 import uk.ac.wellcome.storage.listing.Listing
 import uk.ac.wellcome.storage.listing.s3.S3ObjectLocationListing
 import uk.ac.wellcome.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
 import uk.ac.wellcome.storage.store.s3.S3TypedStore
-import uk.ac.wellcome.storage.tags.Tags
-import uk.ac.wellcome.storage.tags.s3.S3Tags
 import uk.ac.wellcome.storage.transfer.s3.S3PrefixTransfer
 
 class S3ReplicatorTest
@@ -36,8 +36,6 @@ class S3ReplicatorTest
   override def createDstPrefixWith(dstBucket: Bucket): S3ObjectLocationPrefix =
     S3ObjectLocationPrefix(dstBucket.name, keyPrefix = "")
 
-  override val dstTags: Tags[S3ObjectLocation] = new S3Tags()
-
   override val dstStringStore: S3TypedStore[String] = S3TypedStore[String]
 
   override def withPrefixTransfer[R](
@@ -54,4 +52,44 @@ class S3ReplicatorTest
 
   override val dstListing: Listing[S3ObjectLocationPrefix, S3ObjectLocation] =
     S3ObjectLocationListing()
+
+  // The verifier will write a Content-SHA256 checksum tag to objects when it
+  // verifies them.  If an object is then replicated to a new location, any existing
+  // verification tags should be removed.
+  it("doesn't copy tags from the existing objects") {
+    withSrcNamespace { srcNamespace =>
+      withDstNamespace { dstNamespace =>
+        val location = createSrcLocationWith(srcNamespace)
+
+        putSrcObject(location, contents = randomAlphanumeric)
+        srcTags.update(location) { existingTags =>
+          Right(existingTags ++ Map("Content-SHA256" -> "abcdef"))
+        }
+
+        val request = ReplicationRequest(
+          srcPrefix = createSrcPrefixWith(srcNamespace),
+          dstPrefix = createDstPrefixWith(dstNamespace)
+        )
+
+        val result = withReplicator {
+          _.replicate(
+            ingestId = createIngestID,
+            request = request
+          )
+        }
+
+        result shouldBe a[ReplicationSucceeded[_]]
+
+        val dstLocation = createDstLocationWith(
+          dstNamespace,
+          key = location.key
+        )
+
+        srcTags.get(dstLocation).right.value shouldBe Identified(
+          dstLocation,
+          Map.empty
+        )
+      }
+    }
+  }
 }
